# backend/docs/openapi.json
{"openapi":"3.1.0","info":{"title":"다녀감 Backend API","description":"전국 관광지 탐색, 주변 주차장, 익명 세션, GPS 스탬프와 지역 도감 API","contact":{"name":"다녀감 백엔드 팀"},"license":{"name":"Private competition project"},"version":"v1"},"servers":[{"url":"http://127.0.0.1:18080","description":"Generated server url"}],"tags":[{"name":"스탬프","description":"GPS 기반 방문 인증과 스탬프 지급"},{"name":"세션","description":"로그인 전 익명 사용자 세션"},{"name":"주차장","description":"관광지 주변 카카오 PK6 주차장 후보"},{"name":"도감","description":"익명 사용자별 지역 도감과 진행률"},{"name":"관광지","description":"지역, 지도 영역 관광지와 관광지 상세 조회"},{"name":"위치","description":"좌표 기반 주소 조회"}],"paths":{"/api/v1/stamp-verifications":{"post":{"tags":["스탬프"],"summary":"GPS 방문 인증","description":"서버에서 거리·정확도·측정 시각을 검증하고 최초 방문을 원자적으로 저장합니다.","operationId":"verify","parameters":[{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StampVerificationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseStampVerificationResponse"}}}}},"security":[{"anonymousSession":[]}]}},"/api/v1/sessions/anonymous":{"post":{"tags":["세션"],"summary":"익명 세션 생성 또는 재사용","description":"HttpOnly 세션 쿠키를 발급하거나 유효한 기존 세션을 재사용합니다.","operationId":"createOrReuse","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAnonymousSessionResponse"}}}}}}},"/api/v1/geo/reverse":{"post":{"tags":["위치"],"summary":"현재 좌표의 주소 조회","description":"정확한 GPS 좌표를 URL에 남기지 않도록 POST body로 전달합니다.","operationId":"reverse","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReverseGeoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseReverseGeoResponse"}}}}}}},"/api/v1/tourist-spots":{"get":{"tags":["관광지"],"summary":"지도 영역 관광지 조회","description":"화면 bounds 안의 경량 관광지 마커 데이터를 반환합니다.","operationId":"getMapSpots","parameters":[{"name":"northEastLatitude","in":"query","required":true,"schema":{"type":"number"}},{"name":"northEastLongitude","in":"query","required":true,"schema":{"type":"number"}},{"name":"southWestLatitude","in":"query","required":true,"schema":{"type":"number"}},{"name":"southWestLongitude","in":"query","required":true,"schema":{"type":"number"}},{"name":"types","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTouristSpotMapData"}}}}}}},"/api/v1/tourist-spots/{spotId}":{"get":{"tags":["관광지"],"summary":"관광지 상세 조회","description":"상세 정보, 이미지, 방문 상태와 내비 목적지를 반환합니다.","operationId":"getDetail","parameters":[{"name":"spotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTouristSpotDetailResponse"}}}}}}},"/api/v1/tourist-spots/{spotId}/parking":{"get":{"tags":["주차장"],"summary":"관광지 주변 주차장 조회","description":"카카오 장소 검색 결과이며 공영 주차장 여부는 보증하지 않습니다.","operationId":"getParking","parameters":[{"name":"spotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"radiusMeters","in":"query","required":false,"schema":{"type":"integer","format":"int32","minimum":1}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","minimum":1}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseParkingResponse"}}}}}}},"/api/v1/regions":{"get":{"tags":["관광지"],"summary":"지역 목록 조회","description":"활성 광역·시군구 지역 계층을 반환합니다.","operationId":"getRegions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseRegionListData"}}}}}}},"/api/v1/me/collection":{"get":{"tags":["도감"],"summary":"지역 도감 조회","operationId":"getCollection","parameters":[{"name":"regionCode","in":"query","required":true,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","default":"ALL"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseCollectionResponse"}}}}},"security":[{"anonymousSession":[]}]}},"/api/v1/me/collection/summary":{"get":{"tags":["도감"],"summary":"지역별 도감 진행률 조회","operationId":"getSummary","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseCollectionSummaryResponse"}}}}},"security":[{"anonymousSession":[]}]}}},"components":{"schemas":{"GpsPositionRequest":{"type":"object","properties":{"latitude":{"type":"number","maximum":90.0,"minimum":-90.0},"longitude":{"type":"number","maximum":180.0,"minimum":-180.0},"accuracyMeters":{"type":"number","maximum":1000000.0,"minimum":0.0},"measuredAt":{"type":"string","format":"date-time"}},"required":["accuracyMeters","latitude","longitude","measuredAt"]},"StampVerificationRequest":{"type":"object","properties":{"touristSpotId":{"type":"integer","format":"int64"},"position":{"$ref":"#/components/schemas/GpsPositionRequest"}},"required":["position","touristSpotId"]},"ApiMeta":{"type":"object","properties":{"requestId":{"type":"string"},"generatedAt":{"type":"string","format":"date-time"},"count":{"type":"integer","format":"int32"}}},"ApiResponseStampVerificationResponse":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/StampVerificationResponse"},"meta":{"$ref":"#/components/schemas/ApiMeta"}}},"StampVerificationResponse":{"type":"object","properties":{"status":{"type":"string"},"touristSpotId":{"type":"integer","format":"int64"},"distanceMeters":{"type":"number"},"verifiedAt":{"type":"string","format":"date-time"},"visitState":{"type":"string"},"collectionChanged":{"type":"boolean"},"newTitleIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"AnonymousSessionResponse":{"type":"object","properties":{"actorType":{"type":"string"},"expiresAt":{"type":"string","format":"date-time"}}},"ApiResponseAnonymousSessionResponse":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/AnonymousSessionResponse"},"meta":{"$ref":"#/components/schemas/ApiMeta"}}},"GeoPositionRequest":{"type":"object","properties":{"latitude":{"type":"number","maximum":90.0,"minimum":-90.0},"longitude":{"type":"number","maximum":180.0,"minimum":-180.0}},"required":["latitude","longitude"]},"ReverseGeoRequest":{"type":"object","properties":{"position":{"$ref":"#/components/schemas/GeoPositionRequest"}},"required":["position"]},"AddressRegionResponse":{"type":"object","properties":{"depth1":{"type":"string"},"depth2":{"type":"string"},"depth3":{"type":"string"}}},"ApiResponseReverseGeoResponse":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/ReverseGeoResponse"},"meta":{"$ref":"#/components/schemas/ApiMeta"}}},"ReverseGeoResponse":{"type":"object","properties":{"addressName":{"type":"string"},"roadAddressName":{"type":"string"},"region":{"$ref":"#/components/schemas/AddressRegionResponse"},"source":{"type":"string"}}},"ApiResponseTouristSpotMapData":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/TouristSpotMapData"},"meta":{"$ref":"#/components/schemas/ApiMeta"}}},"PositionResponse":{"type":"object","properties":{"latitude":{"type":"number"},"longitude":{"type":"number"}}},"TouristSpotMapData":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/TouristSpotMapItemResponse"}}}},"TouristSpotMapItemResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"position":{"$ref":"#/components/schemas/PositionResponse"},"type":{"type":"string"},"stampEnabled":{"type":"boolean"},"visitState":{"type":"string"},"thumbnailUrl":{"type":"string"}}},"ApiResponseTouristSpotDetailResponse":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/TouristSpotDetailResponse"},"meta":{"$ref":"#/components/schemas/ApiMeta"}}},"NavigationDestinationResponse":{"type":"object","properties":{"destinationName":{"type":"string"},"latitude":{"type":"number"},"longitude":{"type":"number"},"coordinateType":{"type":"string"}}},"TouristSpotAddressResponse":{"type":"object","properties":{"road":{"type":"string"},"lot":{"type":"string"}}},"TouristSpotDetailResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"type":{"type":"string"},"stampEnabled":{"type":"boolean"},"visitState":{"type":"string"},"address":{"$ref":"#/components/schemas/TouristSpotAddressResponse"},"position":{"$ref":"#/components/schemas/PositionResponse"},"overview":{"type":"string"},"images":{"type":"array","items":{"$ref":"#/components/schemas/TouristSpotImageResponse"}},"telephone":{"type":"string"},"homepageUrl":{"type":"string"},"navigation":{"$ref":"#/components/schemas/NavigationDestinationResponse"},"dataSource":{"type":"string"},"dataQuality":{"type":"string"},"lastSyncedAt":{"type":"string","format":"date-time"}}},"TouristSpotImageResponse":{"type":"object","properties":{"url":{"type":"string"},"alt":{"type":"string"},"copyrightType":{"type":"string"}}},"ApiResponseParkingResponse":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/ParkingResponse"},"meta":{"$ref":"#/components/schemas/ApiMeta"}}},"ParkingItemResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"address":{"type":"string"},"position":{"$ref":"#/components/schemas/PositionResponse"},"distanceMeters":{"type":"integer","format":"int32"},"publicVerified":{"type":"boolean"},"source":{"type":"string"},"navigation":{"$ref":"#/components/schemas/NavigationDestinationResponse"},"fetchedAt":{"type":"string","format":"date-time"}}},"ParkingResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/ParkingItemResponse"}},"temporarilyUnavailable":{"type":"boolean"}}},"ApiResponseRegionListData":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/RegionListData"},"meta":{"$ref":"#/components/schemas/ApiMeta"}}},"RegionListData":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/RegionResponse"}}}},"RegionResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"level":{"type":"string"}}},"ApiResponseCollectionResponse":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/CollectionResponse"},"meta":{"$ref":"#/components/schemas/ApiMeta"}}},"CollectionItemResponse":{"type":"object","properties":{"touristSpotId":{"type":"integer","format":"int64"},"name":{"type":"string"},"visitState":{"type":"string"},"verifiedAt":{"type":"string","format":"date-time"},"thumbnailUrl":{"type":"string"}}},"CollectionRegionResponse":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"}}},"CollectionResponse":{"type":"object","properties":{"region":{"$ref":"#/components/schemas/CollectionRegionResponse"},"items":{"type":"array","items":{"$ref":"#/components/schemas/CollectionItemResponse"}}}},"ApiResponseCollectionSummaryResponse":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/CollectionSummaryResponse"},"meta":{"$ref":"#/components/schemas/ApiMeta"}}},"CollectionSummaryItemResponse":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"},"visitedCount":{"type":"integer","format":"int64"},"totalCount":{"type":"integer","format":"int64"},"progressPercent":{"type":"integer","format":"int32"}}},"CollectionSummaryResponse":{"type":"object","properties":{"regions":{"type":"array","items":{"$ref":"#/components/schemas/CollectionSummaryItemResponse"}}}}},"securitySchemes":{"anonymousSession":{"type":"apiKey","description":"POST /api/v1/sessions/anonymous 호출로 발급되는 HttpOnly 익명 세션 쿠키","name":"dg_session","in":"cookie"}}}}