# backend/src/main/java/com/danyeogam/backend/geo/api/ReverseGeoRequest.java
package com.danyeogam.backend.geo.api;

import jakarta.validation.Valid;
import jakarta.validation.constraints.NotNull;

public record ReverseGeoRequest(
        @NotNull @Valid GeoPositionRequest position
) {
}