# backend/src/main/java/com/danyeogam/backend/geo/api/ReverseGeoResponse.java
package com.danyeogam.backend.geo.api;

public record ReverseGeoResponse(
        String addressName,
        String roadAddressName,
        AddressRegionResponse region,
        String source
) {
}